rt_2_wave needs to create multi-track wave file Fixes #139
