+++ examples/rt_2_wave.rs
@@ -1,6 +1,6 @@
-use rtjam_rust::common::box_error::BoxError;
+use log::info;
+use rtjam_rust::{common::box_error::BoxError, server::playback_mixer::PlaybackMixer};
 use clap::Parser;
-use rtjam_rust::server::playback_thread::PlaybackMixer;
 use rtjam_rust::common::get_micro_time;
 use hound;
 
@@ -23,6 +23,8 @@ const FRAME_TIME: u128 = 2_667;
 fn main() -> Result<(), BoxError> {
     let args = Args::parse();
 
+    env_logger::init();
+
     println!("in_file: {}", args.in_file); 
     println!("out_file: {}", args.out_file);
 
@@ -34,9 +36,13 @@ fn main() -> Result<(), BoxError> {
         Err(e) => {dbg!(e);}
     }
 
+    // get the number of channels
+    let ids = mixer.get_ids(now)?;
+    info!("IDs {:?}", ids);
+    let num_chans = ids.len() * 2;
     // Create a wave file
     let spec = hound::WavSpec {
-        channels: 2,
+        channels: num_chans as u16,
         sample_rate: 48000,
         bits_per_sample: 32,
         sample_format: hound::SampleFormat::Float,
@@ -56,13 +62,14 @@ fn main() -> Result<(), BoxError> {
                 mixer.close_stream();
             }            
         }
-        match mixer.get_a_frame() {
+        match mixer.get_mix_channels(num_chans) {
             Some(buf) => {
-                print!(".");
+                // print!(".");
                 for i in 0..buf[0].len() {
-                    // Since file is 2 channel float write left then write (interleave)
-                    writer.write_sample(buf[0][i])?;
-                    writer.write_sample(buf[1][i])?;
+                    //  (interleave)
+                    for j in 0..num_chans {
+                        writer.write_sample(buf[j][i])?;
+                    }
                 }
            }
             None => {
