+++ src/server.rs
@@ -7,3 +7,4 @@ pub mod playback_thread;
 pub mod room_mixer;
 pub mod ping_thread;
 pub mod metronome;
+pub mod playback_mixer;
